requirements.txt
Django==5.0.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pillow==10.3.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
django-cors-headers==4.4.0